Added Login"screen" at the Lighthouse of Blueberry
Fixed a problem, that OnPlayerSpawn would be triggered, when it was
Hooked in the Loginsystem.
Made a command for saveing the camera position in the flymode
filterscript
Changed the function "ReSpawnActor" to "RespawnActor"
Removed some old Code
Added PlayerHooks.inc and saveing the current skin of the player
Added a Jetpack Command
Changed the Location of  some files
Added Virtualworlds for Login
Added an Actor for the Loginscreen - this is more save than using the
Player.
Added an Object an TogglePlayerSpectating the prevent the issue, that
some Objects and the Actor won't stream to the Camera.